JQuery를 이용한 성능 최적화

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -143,11 +143,11 @@ class BookServiceTest @Autowired constructor(
 
         //then
         assertThat(result).hasSize(2)
-        assertCount(result, BookType.COMPUTER, 2)
-        assertCount(result, BookType.SCIENCE, 1)
+        assertCount(result, BookType.COMPUTER, 2L)
+        assertCount(result, BookType.SCIENCE, 1L)
     }
 
-    private fun assertCount(result: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(result: List<BookStatResponse>, type: BookType, count: Long) {
         assertThat(result.first { result -> result.type == type }.count).isEqualTo(count)
     }
 }
